+++ Tests/Functional/Domain/Repository/Registration/RegistrationRepositoryTest.php
@@ -100,6 +100,7 @@ final class RegistrationRepositoryTest extends FunctionalTestCase
         self::assertSame('the internet', $result->getKnownFrom());
         self::assertSame('crash course', $result->getBackgroundKnowledge());
         self::assertTrue($result->hasSeparateBillingAddress());
+        self::assertSame(Registration::ATTENDANCE_MODE_ON_SITE, $result->getAttendanceMode());
     }
 
     /**